modification32

+++ do_it/FormClients.cs
@@ -44,6 +44,7 @@ namespace do_it
         private void activate(Boolean a)
         {
             grpdetails.Enabled = a;
+            password_userTextBox.Enabled = false;
             btncancel.Visible = a;
             btnvalidate.Visible = a;
             btnadd.Visible = !a;
@@ -67,17 +68,21 @@ namespace do_it
 
         private void btnremove_Click(object sender, EventArgs e)
         {
-            SqlConnection cn = new SqlConnection(cs);
-            cn.Open();
-            string req = "delete from users where ID_USER = " + iD_USERTextBox.Text;
-            SqlCommand com = new SqlCommand(req, cn);
-            com.ExecuteNonQuery();
-            usersListBox.Refresh();
-            this.usersTableAdapter.Fill(this.dO_ITDataSet.users);
+            if( MessageBox.Show("Are you sure you want to delete this user ??", "Delete..!", MessageBoxButtons.YesNo) == DialogResult.Yes)
+            {
+                SqlConnection cn = new SqlConnection(cs);
+                cn.Open();
+                string req = "delete from users where ID_USER = " + iD_USERTextBox.Text;
+                SqlCommand com = new SqlCommand(req, cn);
 
-            //usersBindingSource.RemoveCurrent();
-            //tableAdapterManager.UpdateAll(dO_ITDataSet);
-            //activate(true);
+                com.ExecuteNonQuery();
+                usersListBox.Refresh();
+                this.usersTableAdapter.Fill(this.dO_ITDataSet.users);
+
+            }
+            
+
+          
         }
 
         private void btnvalidate_Click(object sender, EventArgs e)
